Partial fix for RFE 741930: Add createElement for elements with 1-for-1 mapping from HTML tag names and DOM Level 0 classes.  This excludes the input tags since the DOM Level 0 classes all map to a single HTML tag.
The change is to add a createElement method in HtmlPage.  To make this work, I added
a document_ member that is set when the page is initialized and can be accessed
during parsing through getDocument.  This internal createElement method is then made
available to JavaScripts by adding jsFunction_createElement to Document.  A test
case was added for a simple case of creating a DIV element.  This partial fix was
noted in the changes.xml file.


git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@303 5f5364db-9458-4db8-a492-e30667be6df6