XMLHttpRequest: Change to more consistent event invocation order
This commit changes the invocation order of events added to XMLHttpRequest,
through both addEventListener() and on-handlers, to be more consistent with
major browser such as Chrome / FF / IE, as well as other events already
implemented in HtmlUnit through EventTarget.

This is done by using the AT_TARGET invocation feature already supplied by
EventTarget in a prior commit which is of a more robust implementation
than the custom implementation in XMLHttpRequest.